refactor : @DiscriminatorColumn 추가
1. @DiscriminatorColumn 추가
2. repository 패키지 생성
