+++ src/main/java/com/flab/shoeauction/domain/user/repository/AdminRepository.java
@@ -1,5 +1,6 @@
-package com.flab.shoeauction.domain.user;
+package com.flab.shoeauction.domain.user.repository;
 
+import com.flab.shoeauction.domain.user.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdminRepository extends JpaRepository<Admin, Long> {
