+++ src/main/java/com/flab/shoeauction/domain/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.flab.shoeauction.domain.user;
+package com.flab.shoeauction.domain.user.repository;
 
+import com.flab.shoeauction.domain.user.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
