+++ src/test/java/com/flab/shoeauction/service/UserServiceTest.java
@@ -13,12 +13,12 @@ import com.flab.shoeauction.controller.dto.AddressBookDto;
 import com.flab.shoeauction.controller.dto.UserDto.ChangePasswordRequest;
 import com.flab.shoeauction.controller.dto.UserDto.FindUserResponse;
 import com.flab.shoeauction.controller.dto.UserDto.SaveRequest;
-import com.flab.shoeauction.domain.AddressBook.Address;
-import com.flab.shoeauction.domain.AddressBook.AddressBook;
-import com.flab.shoeauction.domain.AddressBook.AddressBookRepository;
+import com.flab.shoeauction.domain.addressBook.Address;
+import com.flab.shoeauction.domain.addressBook.AddressBook;
+import com.flab.shoeauction.domain.addressBook.AddressBookRepository;
 import com.flab.shoeauction.domain.user.Account;
 import com.flab.shoeauction.domain.user.User;
-import com.flab.shoeauction.domain.user.UserRepository;
+import com.flab.shoeauction.domain.user.repository.UserRepository;
 import com.flab.shoeauction.exception.user.DuplicateEmailException;
 import com.flab.shoeauction.exception.user.DuplicateNicknameException;
 import com.flab.shoeauction.exception.user.UnableToChangeNicknameException;
